Move lint infrastructure and individual lints into separate files
